Added support xmltv epg to the editor

+++ UtilsLib/rapidxml_utils.hpp
@@ -2,12 +2,15 @@
 #define RAPIDXML_UTILS_HPP_INCLUDED
 
 // Copyright (C) 2006, 2009 Marcin Kalicinski
-// Version 1.13
+// Version 1.14
 // Revision $DateTime: 2009/05/13 01:46:17 $
 //! \file rapidxml_utils.hpp This file contains high-level rapidxml utilities that can be useful
 //! in certain simple scenarios. They should probably not be used if maximizing performance is the main objective.
 
 #include "rapidxml.hpp"
+#include <vector>
+#include <string>
+#include <fstream>
 #include <stdexcept>
 
 namespace rapidxml
@@ -34,7 +37,7 @@ namespace rapidxml
 
             // Determine stream size
             stream.seekg(0, ios::end);
-            size_t size = stream.tellg();
+            size_t size = (size_t)stream.tellg();
             stream.seekg(0);
 
             // Load data and add terminating 0
